Sharoobi_FileShare/backend/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
jinja2==3.1.6
python-multipart==0.0.20
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
redis==6.4.0
PyJWT==2.10.1
passlib==1.7.4
